+++ commons/src/main/java/org/exoplatform/platform/qa/ui/commons/Base.java
@@ -82,10 +82,12 @@ public class Base extends TestBase {
     Screenshots.finishContext();
   }
 
+ /*
   @AfterEach
   public void close() {
     if (!title().equals("Login")) {
       Selenide.close();
     }
   }
+  */
 }
